requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
slack_bolt==1.19.1
slack_sdk==3.31.0
tzlocal==5.2
yarl==1.9.4